KHWE-52: Fix
